+++ demo/src/components/Navbar.tsx
@@ -1,7 +1,7 @@
 import { ThemeToggle } from "./ThemeToggle";
 import styles from "./Navbar.module.scss";
 
-const VERSION = "2.0.1";
+const VERSION = "2.1.0";
 
 export function Navbar() {
   return (