+++ src/types/Country.ts
@@ -1,9 +1,9 @@
-import { CountryCode } from "./CountryCode";
+import { AnyCountryCode } from "./AnyCountryCode";
 
 export type Country = {
   postalCodePatterns: string[];
   examplePostalCodes: string[];
   isGenericRegex: boolean;
   countryName: string;
-  countryCode: CountryCode;
+  countryCode: AnyCountryCode;
 };