+++ src/utils/getAllCountries.ts
@@ -1,9 +1,10 @@
 import { CountryOption } from "../types/CountryOption";
 
-import { COUNTRIES } from "../assets/index";
+import { getActiveData } from "./activeData";
 
 /**
- * Retrieves all countries as an array of CountryOption objects.
+ * Retrieves all countries as an array of CountryOption objects. Includes any
+ * custom countries registered via `configure()`.
  *
  * @returns {CountryOption[]} An array of CountryOption objects, each containing:
  *   - countryName: The name of the country
@@ -18,7 +19,8 @@ import { COUNTRIES } from "../assets/index";
  * // ]
  */
 export const getAllCountries = (): CountryOption[] => {
-  return Object.entries(COUNTRIES).map(([code, country]) => ({
+  const { countries } = getActiveData();
+  return Object.entries(countries).map(([code, country]) => ({
     countryName: country.country,
     countryCode: code.toString(),
   }));